+++ packages/iov-bcp/types/atomicswaptypes.d.ts
@@ -1,7 +1,7 @@
 import { As } from "type-tagger";
 import { Stream } from "xstream";
 import { BcpConnection } from "./connection";
-import { Address, Amount, SwapIdBytes } from "./transactions";
+import { Address, Amount, SwapIdBytes, SwapTimeout } from "./transactions";
 export declare enum SwapState {
     Open = "open",
     Claimed = "claimed",
@@ -18,7 +18,12 @@ export interface SwapData {
      */
     readonly hash: Uint8Array;
     readonly amounts: ReadonlyArray<Amount>;
-    readonly timeout: number;
+    /**
+     * The first point in time at which the offer is expired.
+     *
+     * Can be represented as a block height or UNIX timestamp.
+     */
+    readonly timeout: SwapTimeout;
     readonly memo?: string;
 }
 export interface OpenSwap {